estructura principal de la extension de la api

+++ HazTuHuertoAPI/src/main/java/com/triana/salesianos/HazTuHuertoAPI/controller/VegetableGardenController.java
@@ -86,7 +86,7 @@ public class VegetableGardenController {
     @DeleteMapping("/vegetableGarden/{id}")//Hay que hacer todavia las politicas de borrado
     public ResponseEntity<?> delete(@AuthenticationPrincipal User user,@PathVariable Long id) {
         //Question quest=questionService.findById(id);
-        //if(userService.checkUserLogedInGarden(user.getId(), id))
+        if(userService.checkUserLogedInGarden(user.getId(), id))
             gardenService.deleteById(id);
 
         return ResponseEntity.noContent().build();
